Reorganize TagEditorParam components
