+++ frontend/src/metabase/query_builder/components/template_tags/TagEditorParam.tsx
@@ -1,18 +1,12 @@
-import { Component, useMemo } from "react";
+import { Component } from "react";
 import { connect } from "react-redux";
-import { Link } from "react-router";
 import { t } from "ttag";
 import _ from "underscore";
 
-import Schemas from "metabase/entities/schemas";
-import MetabaseSettings from "metabase/lib/settings";
-import { RequiredParamToggle } from "metabase/parameters/components/RequiredParamToggle";
 import { ValuesSourceSettings } from "metabase/parameters/components/ValuesSourceSettings";
 import type { EmbeddingParameterVisibility } from "metabase/public/lib/types";
-import { SchemaTableAndFieldDataSelector } from "metabase/query_builder/components/DataSelector";
 import { fetchField } from "metabase/redux/metadata";
 import { getMetadata } from "metabase/selectors/metadata";
-import { Flex, Select, Text, TextInputBlurChange } from "metabase/ui";
 import type Database from "metabase-lib/metadata/Database";
 import type Field from "metabase-lib/metadata/Field";
 import type Metadata from "metabase-lib/metadata/Metadata";
@@ -39,12 +33,15 @@ import type { State } from "metabase-types/store";
 
 import {
   ContainerLabel,
-  DefaultParameterValueWidget,
-  ErrorSpan,
   InputContainer,
   TagContainer,
   TagName,
-} from "./TagEditorParam.styled";
+  DefaultRequiredValueControl,
+  FilterWidgetTypeSelect,
+  FieldMappingSelect,
+  FilterWidgetLabelInput,
+} from "./TagEditorParamParts";
+import { VariableTypeSelect } from "./TagEditorParamParts/VariableTypeSelect";
 
 interface Props {
   tag: TemplateTag;
@@ -290,238 +287,3 @@ export const TagEditorParam = connect(
   mapStateToProps,
   mapDispatchToProps,
 )(TagEditorParamInner);
-
-function VariableTypeSelect(props: {
-  value: TemplateTagType;
-  onChange: (value: TemplateTagType) => void;
-}) {
-  return (
-    <InputContainer>
-      <ContainerLabel smallPaddingBottom>{t`Variable type`}</ContainerLabel>
-      <Select
-        value={props.value}
-        placeholder={t`Select…`}
-        onChange={props.onChange}
-        data={[
-          { value: "text", label: t`Text` },
-          { value: "number", label: t`Number` },
-          { value: "date", label: t`Date` },
-          { value: "dimension", label: t`Field Filter` },
-        ]}
-        data-testid="variable-type-select"
-      ></Select>
-    </InputContainer>
-  );
-}
-
-function FieldMappingSelect({
-  tag,
-  hasSelectedDimensionField,
-  table,
-  field,
-  fieldMetadataLoaded,
-  database,
-  databases,
-  setFieldFn,
-}: {
-  tag: TemplateTag;
-  hasSelectedDimensionField: boolean;
-  fieldMetadataLoaded: boolean;
-  table: Table | null | undefined;
-  database?: Database | null;
-  databases: Database[];
-  field: Field | null;
-  setFieldFn: (fieldId: FieldId) => void;
-}) {
-  return (
-    <InputContainer>
-      <ContainerLabel>
-        {t`Field to map to`}
-        {tag.dimension == null && <ErrorSpan>{t`(required)`}</ErrorSpan>}
-      </ContainerLabel>
-
-      {(!hasSelectedDimensionField ||
-        (hasSelectedDimensionField && fieldMetadataLoaded)) && (
-        <Schemas.Loader id={table?.schema?.id}>
-          {() => (
-            <SchemaTableAndFieldDataSelector
-              databases={databases}
-              selectedDatabase={database || null}
-              selectedDatabaseId={database?.id || null}
-              selectedTable={table || null}
-              selectedTableId={table?.id || null}
-              selectedField={field || null}
-              selectedFieldId={
-                hasSelectedDimensionField ? tag?.dimension?.[1] : null
-              }
-              setFieldFn={setFieldFn}
-              className="AdminSelect flex align-center"
-              isInitiallyOpen={!tag.dimension}
-              triggerIconSize={12}
-              renderAsSelect={true}
-            />
-          )}
-        </Schemas.Loader>
-      )}
-    </InputContainer>
-  );
-}
-
-function FilterWidgetTypeSelect({
-  tag,
-  value,
-  onChange,
-  options,
-}: {
-  tag: TemplateTag;
-  value: string;
-  onChange: (widgetType: string) => void;
-  options: { name?: string; type: string }[];
-}) {
-  const hasOptions = options.length > 0;
-  const hasNoWidgetType = tag["widget-type"] === "none" || !tag["widget-type"];
-
-  const optionsOrDefault = useMemo(
-    () =>
-      (hasOptions ? options : [{ name: t`None`, type: "none" }]).map(
-        option => ({
-          label: option.name,
-          value: option.type,
-        }),
-      ),
-    [hasOptions, options],
-  );
-
-  return (
-    <InputContainer>
-      <ContainerLabel smallPaddingBottom>
-        {t`Filter widget type`}
-        {hasNoWidgetType && <ErrorSpan>({t`required`})</ErrorSpan>}
-      </ContainerLabel>
-
-      <Select
-        value={value}
-        onChange={onChange}
-        placeholder={t`Select…`}
-        data={optionsOrDefault}
-        data-testid="filter-widget-type-select"
-      />
-
-      {!hasOptions && (
-        <p>
-          {t`There aren't any filter widgets for this type of field yet.`}{" "}
-          <Link
-            // eslint-disable-next-line no-unconditional-metabase-links-render -- It's hard to tell if this is still used in the app. Please see https://metaboat.slack.com/archives/C505ZNNH4/p1703243785315819
-            to={MetabaseSettings.docsUrl(
-              "questions/native-editor/sql-parameters",
-              "the-field-filter-variable-type",
-            )}
-            target="_blank"
-            className="link"
-          >
-            {t`Learn more`}
-          </Link>
-        </p>
-      )}
-    </InputContainer>
-  );
-}
-
-function FilterWidgetLabelInput({
-  tag,
-  onChange,
-}: {
-  tag: TemplateTag;
-  onChange: (value: string) => void;
-}) {
-  return (
-    <InputContainer>
-      <ContainerLabel>
-        {t`Filter widget label`}
-        {!tag["display-name"] && <ErrorSpan>({t`required`})</ErrorSpan>}
-      </ContainerLabel>
-      <TextInputBlurChange
-        id={`tag-editor-display-name_${tag.id}`}
-        value={tag["display-name"]}
-        onBlurChange={e => onChange(e.target.value)}
-      />
-    </InputContainer>
-  );
-}
-
-function DefaultRequiredValueControl({
-  tag,
-  parameter,
-  isEmbeddedDisabled,
-  onChangeDefaultValue,
-  onChangeRequired,
-}: {
-  tag: TemplateTag;
-  parameter: Parameter;
-  isEmbeddedDisabled: boolean;
-  onChangeDefaultValue: (value: any) => void;
-  onChangeRequired: (value: boolean) => void;
-}) {
-  // We want to remove "default" and "required" so that it
-  // doesn't show up in the default value input update button
-  const parameterAmended = _.omit(
-    tag.type === "text" || tag.type === "dimension"
-      ? parameter || {
-          fields: [],
-          ...tag,
-          type: tag["widget-type"] || null,
-        }
-      : {
-          fields: [],
-          hasVariableTemplateTagTarget: true,
-          type:
-            tag["widget-type"] || (tag.type === "date" ? "date/single" : null),
-        },
-    "default",
-    "required",
-  );
-
-  return (
-    <div>
-      <ContainerLabel>
-        {t`Default filter widget value`}
-        {!tag.default && tag.required && <ErrorSpan>({t`required`})</ErrorSpan>}
-      </ContainerLabel>
-
-      <Flex gap="xs" direction="column">
-        <DefaultParameterValueWidget
-          parameter={parameterAmended}
-          value={tag.default}
-          setValue={onChangeDefaultValue}
-          isEditing
-          commitImmediately
-        />
-
-        <RequiredParamToggle
-          uniqueId={tag.id}
-          disabled={isEmbeddedDisabled}
-          value={tag.required ?? false}
-          onChange={onChangeRequired}
-          disabledTooltip={
-            <>
-              <Text lh={1.4}>
-                {t`This filter is set to disabled in an embedded question.`}
-              </Text>
-              <Text lh={1.4}>
-                {t`To always require a value, first visit embedding settings,
-            make this filter editable or locked, re-publish the
-            question, then return to this page.`}
-              </Text>
-              <Text size="sm">
-                {t`Note`}:{" "}
-                {t`making it locked, will require updating the
-            embedding code before proceeding, otherwise the embed will
-            break.`}
-              </Text>
-            </>
-          }
-        />
-      </Flex>
-    </div>
-  );
-}
