+++ frontend/src/metabase/query_builder/components/template_tags/TagEditorParamParts/index.tsx
@@ -1,0 +1,6 @@
+export { DefaultRequiredValueControl } from "./DefaultRequiredValueControl";
+export { FilterWidgetTypeSelect } from "./FilterWidgetTypeSelect";
+export { FieldMappingSelect } from "./FieldMappingSelect";
+export { VariableTypeSelect } from "./VariableTypeSelect";
+export { FilterWidgetLabelInput } from "./FilterWidgetLabelInput";
+export * from "./TagEditorParam.styled";
